more on add section, still having issues with null rubridto

+++ src/main/java/edu/tcu/cs/peerevaluation/section/SectionService.java
@@ -29,7 +29,7 @@ public class SectionService {
     }
 
     public Section save(Section newSection){
-        newSection.setId(idWorker.nextId());
+        newSection.setId((int)idWorker.nextId());
         return this.sectionRepository.save(newSection);
     }
