+++ src/main/java/edu/tcu/cs/peerevaluation/section/utils/IdWorker.java
@@ -55,7 +55,7 @@ public class IdWorker {
         this.datacenterId = datacenterId;
     }
 
-    public synchronized int nextId() {
+    public synchronized long nextId() {
         long timestamp = timeGen();
         if (timestamp < lastTimestamp) {
             throw new RuntimeException(String.format("Clock moved backwards.  Refusing to generate id for %d milliseconds", lastTimestamp - timestamp));
